로그아웃 비지니스 코드 로직 수정 : 현재 로그인 id 와 refreshToken 추출 id 비교하여 로그아웃

+++ src/main/java/com/example/blogStudy/auth/AuthService.java
@@ -62,14 +62,22 @@ public class AuthService {
 
     // 로그아웃
     @Transactional
-    public void logout(String accessToken, String refreshToken) {
-        try {
-            String userId = jwtProvider.getUserId(refreshToken);
+    public void logout(String userId, String accessToken, String refreshToken) {
+        Claims refreshClaims = jwtProvider.validateRefreshToken(refreshToken);
+        String refreshUserId = refreshClaims.getSubject();
 
-            refreshTokenService.delete(userId);
-            blacklistTokenService.saveBlackList(accessToken);
+        // 현재 로그인한 id 와 각 토큰에서 추출한 id 비교
+        if (!userId.equals(refreshUserId)) {
+            throw new CustomException(ErrorCode.INVALID_TOKEN_OWNER);
         }
-        catch (ExpiredJwtException e) { return; }
+
+        // RefreshToken 유효성 검사
+        if (Boolean.FALSE.equals(refreshTokenService.isValid(userId, refreshToken))) {
+            throw new CustomException(ErrorCode.INVALID_TOKEN);
+        }
+
+        refreshTokenService.delete(userId);
+        blacklistTokenService.saveBlackList(accessToken);
     }
